+++ packages/core/CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - feat(#334): Deprecate `core.mocks.plainRoutes` getter. `core.mock.routes.plain` getter must be used instead.
 - feat(#334): Deprecate `core.mocks.plainRoutesVariants` getter. `core.mock.routes.plainVariants` getter must be used instead.
 - feat(#334): Deprecate `core.mocks.current` getter. `core.mock.collections.selected` getter must be used instead.
+- feat(#334): Deprecate `core.mocks.current` setter. `core.mock.collections.select` method must be used instead.
 
 ## [3.5.0] - 2022-07-05